Update on Skelaton and WeBill and Favorite

+++ Kroya_UI_Project/Screen/Menu/OrderScreen/OrderModel/OrderModel.swift
@@ -7,10 +7,10 @@
 
 import Foundation
 
-// MARK: - Payload
+// MARK: - OrderModel
 struct OrderModel: Identifiable, Decodable {
     var id = UUID()
-    var foodSellId  : Int
+    var foodSellId: Int
     let name: String
     let price: Int?
     let orderCount: Int?
@@ -21,7 +21,7 @@ struct OrderModel: Identifiable, Decodable {
     let foodCardType: String
     let purchaseID: Int?
     let quantity: Int?
-    let totalPrice: Int?
+    let totalPrice: Double?
     let purchaseStatusType: String?
     let purchaseDate: String?
 
@@ -31,28 +31,27 @@ struct OrderModel: Identifiable, Decodable {
         case purchaseID = "purchaseId"
         case quantity, totalPrice, purchaseStatusType, purchaseDate
     }
-    
 }
 
+// MARK: - OrderModelForBuyer
 struct OrderModelForBuyer: Identifiable, Decodable {
     var id: Int // Use `purchaseId` as the unique identifier
     var foodSellId: Int
     var name: String
     var photo: [Photo] // Update to handle an array of photos
     var quantity: Int
-    var totalPrice: Int
+    var totalPrice: Double // Changed from Int to Double
     var dateCooking: String
     var isOrderable: Bool
     var itemType: String
     var foodCardType: String
     var purchaseStatusType: String
     var purchaseDate: String
-  
+
     enum CodingKeys: String, CodingKey {
         case id = "purchaseId"
         case foodSellId, name, photo, quantity, totalPrice, dateCooking, isOrderable, itemType, foodCardType, purchaseStatusType, purchaseDate
     }
 }
-
